requirements.txt
django
djangorestframework
django-filter
psycopg2-binary
gunicorn
pytest
pytest-django
model_bakery
django-environ